requirements.txt
numpy==1.24.3
google-cloud-videointelligence==2.11.0
google-cloud-storage==2.9.0
opencv-python==4.7.0.72
pandas>=1.5.3
matplotlib>=3.7.1
gradio>=3.40.0
pyserial>=3.5